# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GODNSLOG 2.0 API
  description: |
    GODNSLOG 2.0 API - OAST Interaction Verification & Evidence Platform
    
    This API provides endpoints for:
    - Case management (test tasks, vulnerability verification)
    - Payload generation and tracking
    - Interaction capture (DNS, HTTP, SMTP, LDAP, etc.)
    - Evidence chain generation
    - APIKey management with scopes
    - Audit logging
    
    ## Authentication
    Most endpoints require authentication via:
    - JWT Token (user login)
    - APIKey (programmatic access)
    
    ## Version
    Current version: 2.0.0
    
    ## Base URL
    Production: https://api.godnslog.com/v2
    Development: http://localhost:8080/api/v2
  version: 2.0.0
  contact:
    name: GODNSLOG Team
    email: support@godnslog.com
  license:
    name: MIT
    url: https://github.com/chennqqi/godnslog/blob/master/LICENSE
servers:
  - url: http://localhost:8080/api/v2
    description: Development server
  - url: https://api.godnslog.com/v2
    description: Production server
...
paths:
...
  /cases:
    get:
      tags:
        - cases
      summary: List cases
      description: Retrieve list of cases with filtering and pagination
      operationId: listCases
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          description: Items per page
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [active, archived, completed]
        - name: search
          in: query
          description: Search in title and description
          schema:
            type: string
      responses:
        '200':
          description: List of cases
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - cases
      summary: Create case
      description: Create a new case for tracking test tasks
      operationId: createCase
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaseCreateRequest'
      responses:
        '201':
          description: Case created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /payloads:
    get:
      tags:
        - payloads
      summary: List payloads
      description: Retrieve list of payloads with filtering
      operationId: listPayloads
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: case_id
          in: query
          description: Filter by case ID
          schema:
            type: string
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [draft, deployed, hit, archived, expired]
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          description: Items per page
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of payloads
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayloadListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - payloads
      summary: Create payload
      description: Create a new payload with template and variables
      operationId: createPayload
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayloadCreateRequest'
      responses:
        '201':
          description: Payload created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayloadResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /interactions:
    get:
      tags:
        - interactions
      summary: List interactions
      description: Retrieve list of interactions with filtering
      operationId: listInteractions
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: case_id
          in: query
          description: Filter by case ID
          schema:
            type: string
        - name: payload_id
          in: query
          description: Filter by payload ID
          schema:
            type: string
        - name: type
          in: query
          description: Filter by interaction type
          schema:
            type: string
            enum: [dns, http, smtp, ldap, smb, ftp]
        - name: start_time
          in: query
          description: Filter by start time (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          description: Filter by end time (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          description: Items per page
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of interactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - interactions
      summary: Delete interactions
      description: Delete interactions by IDs
      operationId: deleteInteractions
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteRequest'
      responses:
        '204':
          description: Interactions deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /evidence:
    get:
      tags:
        - evidence
      summary: Generate evidence report
      description: Generate evidence chain for a case or payload
      operationId: generateEvidence
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: case_id
          in: query
          description: Case ID
          schema:
            type: string
        - name: payload_id
          in: query
          description: Payload ID
          schema:
            type: string
        - name: format
          in: query
          description: Output format
          schema:
            type: string
            enum: [json, markdown, csv]
            default: json
      responses:
        '200':
          description: Evidence report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /apikeys:
    get:
      tags:
        - apikeys
      summary: List API keys
      description: Retrieve list of API keys for current user
      operationId: listApiKeys
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          description: Items per page
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - apikeys
      summary: Create API key
      description: Create a new API key with specified scopes
      operationId: createApiKey
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyCreateRequest'
      responses:
        '201':
          description: API key created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token from login
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key for programmatic access
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: Error code
        message:
          type: string
          description: Error message
        timestamp:
          type: integer
          format: int64
          description: Unix timestamp
...
    CaseCreateRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: Case title
        description:
          type: string
          description: Case description
        target:
          type: string
          description: Target system or domain
        tags:
          type: array
          items:
            type: string
          description: Case tags
...
    Case:
      type: object
      required:
        - id
        - title
        - created_by
        - created_at
      properties:
        id:
          type: string
          description: Case ID
        title:
          type: string
          description: Case title
        description:
          type: string
          description: Case description
        target:
          type: string
          description: Target system or domain
        status:
          type: string
          enum: [active, archived, completed]
          description: Case status
        tags:
          type: array
          items:
            type: string
        created_by:
          type: string
          description: User ID who created the case
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CaseResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: OK
        data:
          $ref: '#/components/schemas/Case'
    CaseListResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: OK
        data:
          type: object
          required:
            - items
            - total
            - page
            - page_size
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Case'
            total:
              type: integer
            page:
              type: integer
            page_size:
              type: integer
    PayloadCreateRequest:
      type: object
      required:
        - case_id
        - template
      properties:
        case_id:
          type: string
          description: Associated case ID
        template:
          type: string
          description: Payload template (SSRF, XXE, RCE, etc.)
        variables:
          type: object
          additionalProperties:
            type: string
          description: Custom variables for template rendering
        expires_at:
          type: string
          format: date-time
          description: Payload expiration time
        expected_protocol:
          type: string
          enum: [dns, http, smtp, ldap]
          description: Expected interaction protocol
    Payload:
      type: object
      required:
        - id
        - case_id
        - token
        - template
        - created_by
        - created_at
      properties:
        id:
          type: string
          description: Payload ID
        case_id:
          type: string
          description: Associated case ID
        token:
          type: string
          description: Unique tracking token
        template:
          type: string
          description: Payload template type
        rendered_payload:
          type: string
          description: Rendered payload with variables
        variables:
          type: object
          additionalProperties:
            type: string
        status:
          type: string
          enum: [draft, deployed, hit, archived, expired]
        expected_protocol:
          type: string
          enum: [dns, http, smtp, ldap]
        expires_at:
          type: string
          format: date-time
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PayloadResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: OK
        data:
          $ref: '#/components/schemas/Payload'
    PayloadListResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: OK
        data:
          type: object
          required:
            - items
            - total
            - page
            - page_size
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Payload'
            total:
              type: integer
            page:
              type: integer
            page_size:
              type: integer
    Interaction:
      type: object
      required:
        - id
        - type
        - timestamp
        - source_ip
      properties:
        id:
          type: string
          description: Interaction ID
        type:
          type: string
          enum: [dns, http, smtp, ldap, smb, ftp]
        case_id:
          type: string
        payload_id:
          type: string
        token:
          type: string
        timestamp:
          type: string
          format: date-time
        source_ip:
          type: string
        # DNS specific
        domain:
          type: string
        dns_type:
          type: string
        # HTTP specific
        method:
          type: string
        path:
          type: string
        headers:
          type: object
          additionalProperties:
            type: string
        body:
          type: string
        user_agent:
          type: string
        content_type:
          type: string
        # Common
        raw_data:
          type: string
          description: Raw interaction data
...
    InteractionListResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: OK
        data:
          type: object
          required:
            - items
            - total
            - page
            - page_size
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Interaction'
            total:
              type: integer
            page:
              type: integer
            page_size:
              type: integer
    DeleteRequest:
      type: object
      required:
        - ids
      properties:
        ids:
          type: array
          items:
            type: string
          description: List of IDs to delete
...
    EvidenceResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: OK
        data:
          type: object
          properties:
            format:
              type: string
            content:
              type: string
            metadata:
              type: object
    ApiKeyCreateRequest:
      type: object
      required:
        - name
        - scopes
      properties:
        name:
          type: string
          description: API key name
        scopes:
          type: array
          items:
            type: string
          description: API key scopes
        expires_at:
          type: string
          format: date-time
          description: Expiration time
    ApiKey:
      type: object
      required:
        - id
        - key
        - name
        - scopes
        - created_by
        - created_at
      properties:
        id:
          type: string
        key:
          type: string
          description: API key value (only shown on creation)
        key_prefix:
          type: string
          description: API key prefix for identification
        name:
          type: string
        scopes:
          type: array
          items:
            type: string
        expires_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
    ApiKeyResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: OK
        data:
          $ref: '#/components/schemas/ApiKey'
    ApiKeyListResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: OK
        data:
          type: object
          required:
            - items
            - total
            - page
            - page_size
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/ApiKey'
            total:
              type: integer
            page:
              type: integer
            page_size:
              type: integer
...